Web/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Computer Vision & ML
opencv-python==4.8.1.78
numpy==1.23.5
tensorflow==2.12.0
pillow==10.0.1

# Utilities
click
cryptography